[feat] 테이블 헤더 컴포넌트 분리 #179
