+++ src/constants/tooltips.ts
@@ -1,23 +1,14 @@
 export const TTS_TOOLTIP = {
-  APPLY: '선택한 항목에 TTS 설정 적용하기',
-  RESET_SETTINGS: 'TTS 설정 변경사항 초기화하기',
   REGENERATE_SELECTED: '선택한 텍스트 재생성하기',
   DOWNLOAD_AUDIO: '선택한 음성 파일 다운로드',
   VIEW_HISTORY: '음성 생성 기록 및 히스토리 보기',
 } as const;
 
 export const VC_TOOLTIP = {
-  APPLY: '선택한 항목에 VC 설정 적용하기',
-  RESET_SETTINGS: 'VC 설정 변경사항 초기화하기',
   UPLOAD_TEXT: {
-    ENABLED: '선택된 오디오 파일의 설명 텍스트를 업로드합니다',
+    ENABLED: '선택된 오디오 파일의 스크립트를 업로드해주세요',
     DISABLED: '먼저 오디오 파일을 업로드해주세요',
   },
-} as const;
-
-export const CONCAT_TOOLTIP = {
-  APPLY: '선택한 항목에 CONCAT 설정 적용하기',
-  RESET_SETTINGS: 'CONCAT 설정 변경사항 초기화하기',
 } as const;
 
 export const AUDIO_FOOTER_TOOLTIP = {
