+++ src/types/table.ts
@@ -39,3 +39,23 @@ export interface ConcatTableItem extends TableItem {
   fileName: string;
   silentRegions: { start: number; end: number }[];
 }
+
+export interface ListRowProps {
+  id: string;
+  text: string;
+  isSelected: boolean;
+  onPlay: (id: string) => void;
+  onSelectionChange: (id: string) => void;
+  onTextChange: (id: string, newText: string) => void;
+  type?: 'TTS' | 'VC' | 'CONCAT';
+  fileName?: string;
+  speed: number;
+  volume: number;
+  pitch: number;
+  convertedAudioUrl?: string;
+  status?: '대기중' | '완료' | '실패' | '진행';
+  targetVoice?: string;
+  frontSilence?: number;
+  backSilence?: number;
+  endSilence?: number;
+}
